requirements.txt
Django==5.1.4
python-pptx== 1.0.2
matplotlib==3.9.3
requests== 2.32.3
Pillow==11.0.0
django-rest-framework==3.15.2
django-rest-framework-simplejwt==5.3.1
django-environ==0.11.2
drf-yasg==1.21.8
openai==1.56.2
gunicorn==23.0.0
redis==5.2.0
nginxpy==0.1.0
black==24.10.0
pre-commit==
ruff==0.8.2
